[NAYB-20] refactor: `UserRepository` 반환 타입인 `UserOrderCount`의 위치를 이동
`user/repository/response` 하위로 이동시켰음
