+++ src/main/java/com/prgrms/nabmart/domain/user/repository/response/UserOrderCount.java
@@ -1,4 +1,4 @@
-package com.prgrms.nabmart.domain.user.repository;
+package com.prgrms.nabmart.domain.user.repository.response;
 
 public interface UserOrderCount {
     Long getUserId();
